code/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
mangum==0.18.0
python-dotenv==1.0.1
